clean mlx qwen_vl code

+++ tllm/__init__.py
@@ -34,7 +34,7 @@ else:
     import mlx.core as mx
     import numpy as np
 
-    DTYPE = mx.float16
+    DTYPE = mx.bfloat16
     CONVERT_DTYPE = mx.float16
     DES_DTYPE = np.float16
     DEVICE = None
